statesvc: clamp legacy MaxValueBytes once in the index; regen swagger doc; drop a redundant conversion
- FunctionIndex.Upsert clamps a stored MaxValueBytes to the admission
  ceiling, so the KV PUT/CAS routes and the EventLog routes (and the body
  caps derived from the quota) enforce one bound on both the bearer and the
  admin path; the EventLog-only clamp in the handler goes away. A Function
  stored before the ceiling existed can no longer admit a value the admin
  HMAC path would reject.
- Regenerate the swagger type doc for the MaxValueBytes comment change.
- Remove the int64 conversions golangci-lint (unconvert) flagged in the
  embedded transport cap test.

+++ pkg/statestore/httpapi/limits_test.go
@@ -20,5 +20,5 @@ import (
 // append or KV write it already admitted.
 func TestMaxRequestBytesCoversStatesvc(t *testing.T) {
 	t.Parallel()
-	assert.GreaterOrEqual(t, int64(httpapi.MaxRequestBytes), int64(stateapi.MaxRequestBodyBytes))
+	assert.GreaterOrEqual(t, httpapi.MaxRequestBytes, stateapi.MaxRequestBodyBytes)
 }
